rename AnswerQuestion package for snake_case consistency
