+++ src/use_case/answer_question/AnswerQuestionInputBoundary.java
@@ -1,4 +1,4 @@
-package use_case.AnswerQuestion;
+package use_case.answer_question;
 
 public interface AnswerQuestionInputBoundary {
     void answer(AnswerQuestionInputData inputData);
