+++ src/use_case/answer_question/AnswerQuestionInteractor.java
@@ -1,4 +1,4 @@
-package use_case.AnswerQuestion;
+package use_case.answer_question;
 
 import entity.Question;
 import entity.Quiz;
